fix(backend): NUMERIC 컬럼 엔티티 타입 정합성 + application.yml 중복 키 제거
- monthly_sales_history.total_ledger_amount, product.pallet 가 V17에서 NUMERIC(18,4) 로 정의됐으나 엔티티가 Double 이라 Hibernate schema validate 가 부팅을 차단하던 문제 수정 (#575 후행 보정)
- 엔티티 필드를 BigDecimal? 로 변경하고 @Column 에 precision/scale 명시
- SapMonthlySalesHistoryService / SapProductMasterService 의 set 직전에 BigDecimal.valueOf 로 변환 (parseAmount/parseDouble 흐름은 그대로)
- 관련 테스트 isEqualByComparingTo(BigDecimal) 로 변경
- application.yml 의 중복 'app:' 루트 키 제거 (#582 머지 잔여물 추정)

+++ backend/src/main/kotlin/com/otoki/powersales/sap/inbound/service/SapMonthlySalesHistoryService.kt
@@ -18,6 +18,7 @@ import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Service
 import org.springframework.web.context.request.RequestContextHolder
 import org.springframework.web.context.request.ServletRequestAttributes
+import java.math.BigDecimal
 
 /**
  * SAP 월 매출 이력 인바운드 UPSERT 서비스. (Spec #560)
@@ -152,7 +153,7 @@ class SapMonthlySalesHistoryService(
         entity.abcClosingAmount1 = amounts[0]
         entity.abcClosingAmount2 = amounts[1]
         entity.abcClosingAmount3 = amounts[2]
-        entity.totalLedgerAmount = amounts[3] // Spec #575: SAP TotalLedgerAmount 보존
+        entity.totalLedgerAmount = BigDecimal.valueOf(amounts[3]) // Spec #575: SAP TotalLedgerAmount 보존
         entity.shipClosingAmount = amounts[4]
         entity.rlsalesC = amounts[5]
     }
